+++ docs/changes.rst
@@ -11,6 +11,7 @@ Weblate 5.13.1
 
 * Access control for :http:get:`/api/users/(str:username)/`.
 * :ref:`file_format_params` were not properly applied in some situations.
+* :ref:`mt-libretranslate` compatibility with LibreTranslate 1.7.0.
 
 .. rubric:: Compatibility